pyproject.toml
[project]
name = "fittrack-etl-pipeline"
version = "1.0.0"
description = "Production-grade ETL pipeline for FitTrack Pro fitness data"
requires-python = ">=3.10"
authors = [
    { name = "Clement Ohenhen Jr", email = "clement@example.com" }
]

dependencies = [
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "apscheduler>=3.10.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]
dashboard = [
    "streamlit>=1.28.0",
]

[project.scripts]
fittrack-pipeline = "src.orchestrator:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
